Removed toString/read law from Read
Read should work together with show not with toString. The reason
for this change is, that sometimes it is not possible to
implement toString consistently with the Show instance since
in order to implement Show, another type class must be present.

This happened for instance for class DMap in the RpgTool, were
a Show instance can be derived if the key type has an RpgEnum
instance and the value type an instance of Show. In this case
it was impossible to implement toString in a sane manner. This
shows how broken the concept of toString is.